massively satisfying refactor
