+++ src/App/Auth/index.js
@@ -1,0 +1,9 @@
+import Login from './Login'
+import OAuth from './OAuth'
+import PrivateRoute from './PrivateRoute'
+
+export {
+  Login,
+  OAuth,
+  PrivateRoute
+}
