+++ src/App/Playlist/MediaPlayer/Controls/Controls.scss
