+++ src/App/Playlist/MediaPlayer/MediaPlayer.scss
