+++ src/App/Playlist/PlaylistHeader/PlaylistHeader.scss
