+++ src/App/Playlist/Tracklist/TrackContainer/AlbumCover/AlbumCover.scss
