+++ src/App/Playlist/Tracklist/TrackContainer/AlbumCover/index.js
