+++ src/App/Playlist/Tracklist/TrackContainer/MainInfo/MainInfo.scss
