+++ src/App/Playlist/Tracklist/TrackContainer/ProgressBar/ProgressBar.scss
